+++ app/src/main/java/com/example/t_bank/data/model/BudgetRequest.kt
@@ -4,6 +4,6 @@ import com.google.gson.annotations.SerializedName
 
 data class BudgetRequest(
     @SerializedName("userId") val userId: Long,
-    @SerializedName("income") val income: Double,
+    @SerializedName("income") val income: Int,
     @SerializedName("categories") val categories: List<BudgetCategory>
 )
